Created image inser hook
